arwa/requirements.txt
asgiref==3.7.2
attrs==23.2.0
colorama==0.4.6
Django==5.0
django-js-asset==2.2.0
django-mptt==0.15.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==22.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
